requirements.txt
annotated-types==0.7.0
anyio==4.9.0
appdirs==1.4.4
asciitree==0.3.3
attrs==25.3.0
boto3==1.37.28
botocore==1.37.28
cattrs==24.1.3
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.1
contourpy==1.3.1
copernicusmarine==2.0.1
cycler==0.12.1
dask==2025.3.0
Deprecated==1.2.18
fasteners==0.19
fonttools==4.57.0
fsspec==2025.3.2
geojson-rewind==1.1.0
geomet==1.1.0
h11==0.14.0
h5netcdf==1.6.1
h5py==3.13.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jmespath==1.0.1
kiwisolver==1.4.8
locket==1.0.0
lxml==5.3.2
matplotlib==3.10.1
numcodecs==0.15.1
numpy==2.2.4
packaging==24.2
pandas==2.2.3
partd==1.4.2
pillow==11.1.0
platformdirs==4.3.7
pydantic==2.11.2
pydantic_core==2.33.1
pygbif==0.6.5
pyparsing==3.2.3
pystac==1.12.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
requests-cache==1.2.1
s3transfer==0.11.4
semver==3.0.4
setuptools==78.1.0
six==1.17.0
sniffio==1.3.1
toolz==1.0.0
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.1
tzdata==2025.2
url-normalize==2.2.0
urllib3==2.3.0
wrapt==1.17.2
xarray==2025.3.1
zarr==2.18.5
